requirements.txt
streamlit>=1.28.0
networkx>=3.0.0
pyvis>=0.3.0
pandas>=2.0.0
numpy>=1.21.0
scipy
plotly
node2vec
scikit-learn
gensim
chromadb>=0.4.0
sentence-transformers>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
tqdm>=4.64.0
